Refactor the mess
