+++ physics-be/server/config.py
@@ -2,5 +2,5 @@ import os
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
-SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'app.db')
+SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'server.db')
 SQLALCHEMY_MIGRATE_REPO = os.path.join(basedir, 'migrations')
